Fix synthia evaluator

+++ research/object_detection/eval_active_learner_synthia.py
@@ -180,23 +180,7 @@ if __name__ == "__main__":
             graph_rewriter_fn = graph_rewriter_builder.build(
                 configs['graph_rewriter_config'], is_training=False)
 
-       # # Need to reset graph for evaluation
-        #tf.reset_default_graph()
-
-        #metrics,_,_ = evaluator.evaluate(
-          #create_eval_input_dict_fn,
-          #eval_model_fn,
-          #eval_config,
-          #categories,
-          #train_dir,
-          #eval_dir,
-          #graph_hook_fn=graph_rewriter_fn)
-
-
-        ## Done with previous cycle
-        #if os.path.exists(future_train_dir):
-
-        # Make sure we save the very last one, evaluate one last time
+       # Need to reset graph for evaluation
         tf.reset_default_graph()
 
         metrics,_,_ = evaluator.evaluate(
@@ -208,20 +192,35 @@ if __name__ == "__main__":
           eval_dir,
           graph_hook_fn=graph_rewriter_fn)
 
-        #aps = [metrics[keyAll],[metrics[keyBike], metrics[keyCar],metrics[keyMotorbike]]]
-        aps = [metrics[keyAll]]
+
+        ## Done with previous cycle
+        if os.path.exists(future_train_dir):
+
+            # Make sure we save the very last one, evaluate one last time
+            tf.reset_default_graph()
+
+            metrics,_,_ = evaluator.evaluate(
+              create_eval_input_dict_fn,
+              eval_model_fn,
+              eval_config,
+              categories,
+              train_dir,
+              eval_dir,
+              graph_hook_fn=graph_rewriter_fn)
+
+            #aps = [metrics[keyAll],[metrics[keyBike], metrics[keyCar],metrics[keyMotorbike]]]
+            aps = [metrics[keyAll]]
 
 
-        performances['R'+str(run_num)+'c'+str(cycle)]= aps
+            performances['R'+str(run_num)+'c'+str(cycle)]= aps
 
-        # Write current performance
-        json_str = json.dumps(performances)
-        f = open(output_file,'w')
-        f.write(json_str)
-        f.close()
+            # Write current performance
+            json_str = json.dumps(performances)
+            f = open(output_file,'w')
+            f.write(json_str)
+            f.close()
 
-        #cycle +=1
-        #train_dir = future_train_dir
-        #future_train_dir = FLAGS.train_dir + name + 'R' + str(run_num) + 'cycle' + str(cycle+1) + '/'
-
+            cycle +=1
+            train_dir = future_train_dir
+            future_train_dir = FLAGS.train_dir + name + 'R' + str(run_num) + 'cycle' + str(cycle+1) + '/'
